requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
django-picklefield==3.4.0
django-q2==1.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
whitenoise==6.12.0
